requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
faiss-cpu
altair
